[#13] Create printf and gnl dir in lib/ and move .c files to each directory
