+++ libs/libft/gnl/get_next_line.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/libft.h"
+#include "../../include/libft.h"
 
 static char	*create_newline_frm_save(char *save, bool is_include_nl)
 {
