+++ libs/libft/gnl/get_next_line_utils.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/libft.h"
+#include "../../include/libft.h"
 
 size_t	ft_strlen_gnl(const char *str)
 {
