+++ libs/libft/printf/ft_printf.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/libft.h"
+#include "../../include/libft.h"
 
 static t_printf_info	*init_info(int fd)
 {
