+++ libs/libft/printf/ft_put_for_printf.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/libft.h"
+#include "../../include/libft.h"
 
 ssize_t	ft_putchar_for_printf(char c, int fd)
 {
