+++ libs/libft/printf/input_printf_format.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/libft.h"
+#include "../../include/libft.h"
 
 void	get_flag(char *fmt, t_printf_info *info)
 {
