+++ libs/libft/printf/print_num.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/libft.h"
+#include "../../include/libft.h"
 
 static char	*convert_u2base(unsigned long u, int base, int capitals)
 {
