+++ libs/libft/printf/print_string.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/libft.h"
+#include "../../include/libft.h"
 
 ssize_t	print_c(char chr, t_printf_info info)
 {
