+++ libs/libft/printf/printf_controller.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/libft.h"
+#include "../../include/libft.h"
 
 static void	init_info_for_fmt(t_printf_info *info)
 {
