+++ libs/libft/printf/vaild_info_for_each_fmt.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/libft.h"
+#include "../../include/libft.h"
 
 static int	is_valid_strs(t_printf_info info, char type)
 {
